Connection: add new method to fetch a connected site's blog ID.
This should avoid duplication of code in different parts of our codebase.
