+++ projects/packages/connection/changelog/add-shared-get-blog-id
@@ -1,0 +1,4 @@
+Significance: minor
+Type: added
+
+Add new method to get a connected site's blog ID.
