+++ projects/plugins/jetpack/_inc/lib/class-jetpack-tweetstorm-helper.php
@@ -7,6 +7,7 @@
  */
 
 use Automattic\Jetpack\Connection\Client;
+use Automattic\Jetpack\Connection\Utils as Connection_Utils;
 use Automattic\Jetpack\Status;
 use Twitter\Text\Regex as Twitter_Regex;
 use Twitter\Text\Validator as Twitter_Validator;
@@ -262,7 +263,7 @@ class Jetpack_Tweetstorm_Helper {
 			);
 		}
 
-		$site_id = self::get_site_id();
+		$site_id = Connection_Utils::get_site_id();
 		if ( is_wp_error( $site_id ) ) {
 			return $site_id;
 		}
@@ -1732,23 +1733,5 @@ class Jetpack_Tweetstorm_Helper {
 		if ( ! wp_http_validate_url( $redirect_url ) ) {
 			throw new Requests_Exception( __( 'A valid URL was not provided.', 'jetpack' ), 'wp_http.redirect_failed_validation' );
 		}
-	}
-
-	/**
-	 * Get the WPCOM or self-hosted site ID.
-	 *
-	 * @return mixed
-	 */
-	public static function get_site_id() {
-		$is_wpcom = ( defined( 'IS_WPCOM' ) && IS_WPCOM );
-		$site_id  = $is_wpcom ? get_current_blog_id() : Jetpack_Options::get_option( 'id' );
-		if ( ! $site_id ) {
-			return new WP_Error(
-				'unavailable_site_id',
-				__( 'Sorry, something is wrong with your Jetpack connection.', 'jetpack' ),
-				403
-			);
-		}
-		return (int) $site_id;
 	}
 }
