+++ projects/plugins/jetpack/changelog/add-shared-get-blog-id
@@ -1,0 +1,4 @@
+Significance: patch
+Type: other
+
+General: switch to shared method for getting a WordPress.com blog ID.
